cr_tests/eclone: fix -Wall warnings
eclone-1.c: In function 'do_eclone':
eclone-1.c:56: warning: cast from pointer to integer of different size
eclone-1.c:58: warning: cast from pointer to integer of different size
eclone-1.c:59: warning: cast from pointer to integer of different size

[and similar for eclone-[2-5] ]

Cast pointers to unsigned long, not u64.  Remove unnecessary (u64)0
casts.

+++ eclone/eclone-2.c
@@ -43,10 +43,10 @@ static int do_eclone(int (*child_fn)(void *), void *child_arg,
 	}
 
 	memset(&clone_args, 0, sizeof(clone_args));
-	clone_args.child_stack = (u64)genstack_sp(stack);
-	clone_args.child_stack_size = (u64)0;
-	clone_args.parent_tid_ptr = (u64)(&parent_tid);
-	clone_args.child_tid_ptr = (u64)(&child_tid);
+	clone_args.child_stack = (unsigned long)genstack_sp(stack);
+	clone_args.child_stack_size = 0;
+	clone_args.parent_tid_ptr = (unsigned long)(&parent_tid);
+	clone_args.child_tid_ptr = (unsigned long)(&child_tid);
 	clone_args.nr_pids = nr_pids;
 
 	if (verbose) {
